% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_amendments_bill.R
\name{get_amendments_bill}
\alias{get_amendments_bill}
\title{Get Amendments for a Specific Bill}
\usage{
get_amendments_bill(congress, bill_id, myAPI_Key)
}
\arguments{
\item{congress}{105-115@param myAPI_Key  use the congress API, you must sign up for an API key. The API key must be included in all API requests to the server, set as a header.}

\item{bill_id}{a bill slug, for example hr4881 - these can be found in the recent bill response.}

\item{myAPI_Key}{use the congress API, you must sign up for an API key. The API key must be included in all API requests to the server, set as a header.}
}
\value{
List of returned JSON from endpoint that retrieves the 20 most recent results and supports paginated requests.
}
\description{
HTTP Request: GET https://api.propublica.org/congress/v1/{congress}/bills/{bill-id}/amendments.json. Use this request type to get Library of Congress-assigned subjects about a particular bill.
}
\examples{
\dontrun{
get_amendments_bill(115, 'hr1628')
}
}
