% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_committee.R
\name{get_committee}
\alias{get_committee}
\title{Get Committee Info
HTTP Request: GET https://api.propublica.org/campaign-finance/v1/{cycle}/committees/{fec-id}}
\usage{
get_committee(FEC_ID, cycle = 2018, page = 1, myAPI_Key)
}
\arguments{
\item{FEC_ID}{The FEC-assigned 9-character ID of a committee.}

\item{cycle}{The election cycle}

\item{page}{Results come in batches of 20. The page parameter specifies which one to retrieve.}

\item{myAPI_Key}{API key from Propublica To use the congress API, you must sign up for an API key.}
}
\value{
List of returned JSON from endpoint that retrieves ca specific FEC committee for a given campaign cycle from the campaign-finance API from ProPublica.
}
\description{
Get Committee Info
HTTP Request: GET https://api.propublica.org/campaign-finance/v1/{cycle}/committees/{fec-id}
}
\examples{
\donttest{
get_committee('C00553560', 2016)
}
}
