% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_committee_by_name.R
\name{get_committee_by_name}
\alias{get_committee_by_name}
\title{Get Committee by name}
\usage{
get_committee_by_name(NAME, cycle = 2018, page = 1, myAPI_Key)
}
\arguments{
\item{NAME}{The name of the committee.}

\item{cycle}{The election cycle}

\item{page}{Results come in batches of 20. The page parameter specifies which one to retrieve.}

\item{myAPI_Key}{API key from Propublica To use the congress API, you must sign up for an API key.}
}
\value{
List of returned JSON from endpoint that retrieves ca specific committee by name for a given campaign cycle from the campaign-finance API from ProPublica.
}
\description{
HTTP Request: GET https://api.propublica.org/campaign-finance/v1/{cycle}/committees/search
}
\examples{
\donttest{
get_committee_by_name('Americans FOR A BETTER TOMORROW TODAY', 2016)
}
}
