% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_committee_filings.R
\name{get_committee_filings}
\alias{get_committee_filings}
\title{Get Committee Filings}
\usage{
get_committee_filings(FEC_ID, cycle = 2018, myAPI_Key)
}
\arguments{
\item{FEC_ID}{The FEC-assigned 9-character ID of a committee.}

\item{cycle}{The election cycle}
}
\description{
HTTP Request GET https://api.propublica.org/campaign-finance/v1/{cycle}/committees/{fec-id}/filings
}
\examples{
\dontrun{
get_committee_filings('C00553560', 2016)
}
}
