% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_committee_leadership.R
\name{get_committee_leadership}
\alias{get_committee_leadership}
\title{Get Leadership Committees}
\usage{
get_committee_leadership(cycle = 2018, page = 1, myAPI_Key)
}
\arguments{
\item{cycle}{The election cycle: an even-numbered year between 1996 and 2016. The current cycle in 2018}

\item{page}{Results come in batches of 20. The page parameter specifies which one to retrieve.}

\item{myAPI_Key}{API key from Propublica To use the congress API, you must sign up for an API key.}
}
\value{
List of returned JSON from endpoint that retrieves committees designated as “leadership PACs” by the FEC from the campaign-finance API from ProPublica.
}
\description{
HTTP Request: GET https://api.propublica.org/campaign-finance/v1/{cycle}/committees/leadership
}
\examples{
\donttest{
get_committee_leadership(2016)
}
}
