% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/get_independent_expenditure_support_presidential_candidate.R
\name{get_independent_expenditure_support_presidential_candidate}
\alias{get_independent_expenditure_support_presidential_candidate}
\title{Get Independent Expenditures that Support or Oppose Presidential Candidates}
\usage{
get_independent_expenditure_support_presidential_candidate(cycle = 2018,
  myAPI_Key)
}
\arguments{
\item{cycle}{The election cycle}

\item{myAPI_Key}{use the Campaign Finance API, you must sign up for an API key. The API key must be included in all API requests to the server, set as a header.}
}
\value{
List of returned JSON from endpoint that retrieves the 200 most recent independent expenditures in support of or opposition to any presidential candidate.
}
\description{
#HTTP Request: GET https://api.propublica.org/campaign-finance/v1/{cycle}/president/independent_expenditures
}
\examples{
\dontrun{
get_independent_expenditure_support_presidential_candidate(2016)
}
}
