% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_races_for_state.R
\name{get_races_for_state}
\alias{get_races_for_state}
\title{Get all Candidates in a State for an Election Cycle}
\usage{
get_races_for_state(state, cycle = 2018)
}
\arguments{
\item{state}{The State (US) of the election}

\item{cycle}{Which election cycle (e.g. 2018)}
}
\value{
the result of the API call.
}
\description{
Get all Candidates in a State for an Election Cycle
}
\examples{
\dontrun{
get_races_for_state('DE')
get_races_for_state('DE', 2016)
}
}
