% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recent_committee_hearing.R
\name{get_recent_committee_hearing}
\alias{get_recent_committee_hearing}
\title{Get Recent Committee Hearings}
\usage{
get_recent_committee_hearing(congress, page = 1, myAPI_Key)
}
\arguments{
\item{congress}{114-116}

\item{page}{Results come in batches of 20. The page parameter specifies which one to retrieve.}

\item{myAPI_Key}{API key from Propublica To use the congress API, you must sign up for an API key.}
}
\value{
List of returned JSON from endpoint that retrieves the 20 most recent results and supports pagination using multiples of 20.
}
\description{
To get a list of 20 upcoming Senate or House committee meetings, use the following function. HTTP Request: GET https://api.propublica.org/congress/v1/{congress}/committees/hearings.json
}
\examples{
\donttest{
get_recent_committee_hearing(115)
}
}
