% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recent_congressional_statements.R
\name{get_recent_congressional_statements}
\alias{get_recent_congressional_statements}
\title{Get Recent Congressional Statements}
\usage{
get_recent_congressional_statements(page = 1, myAPI_Key)
}
\arguments{
\item{page}{Results come in batches of 20. The page parameter specifies which one to retrieve.}

\item{myAPI_Key}{API key from Propublica To use the congress API, you must sign up for an API key.}
}
\value{
List of returned JSON from endpoint that retrieves  the 20 most recent results and supports pagination using multiples of 20.
}
\description{
To get lists of recent statements published on congressional websites, use the following function. HTTP Request: GET https://api.propublica.org/congress/v1/statements/latest.json
}
\examples{
\donttest{
get_recent_congressional_statements()
}
}
