% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recent_late_contributions.R
\name{get_recent_late_contributions}
\alias{get_recent_late_contributions}
\title{Get Recent Late Contributions}
\usage{
get_recent_late_contributions(cycle = 2018, myAPI_Key)
}
\arguments{
\item{cycle}{The election cycle}

\item{myAPI_Key}{To use the Campaign Finance API, you must sign up for an API key. The API key must be included in all API requests to the server, set as a header.}
}
\value{
List of returned JSON from endpoint that retrieves the most recent late contributions to candidates from the campaign-finance API from ProPublica.
}
\description{
HTTP Request: GET https://api.propublica.org/campaign-finance/v1/{cycle}/contributions/48hour
}
\examples{
\dontrun{
get_recent_late_contributions(2016)
}
}
