% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recent_late_contributions_committee.R
\name{get_recent_late_contributions_committee}
\alias{get_recent_late_contributions_committee}
\title{Get Recent Late Contributions to a Specific Committee}
\usage{
get_recent_late_contributions_committee(cycle = 2018, FEC_ID, myAPI_Key)
}
\arguments{
\item{cycle}{The election cycle}

\item{FEC_ID}{The FEC-assigned 9-character ID of a committee.}

\item{myAPI_Key}{To use the Campaign Finance API, you must sign up for an API key. The API key must be included in all API requests to the server, set as a header.}
}
\description{
HTTP Request: GET https://api.propublica.org/campaign-finance/v1/{cycle}/committees/{fec-id}/48hour
}
\examples{
\dontrun{
get_recent_late_contributions_committee(2016, 'C00575050')
}
}
