% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recent_late_contributions_date(issue).R
\name{get_recent_late_contributions_date}
\alias{get_recent_late_contributions_date}
\title{Get Recent Late Contributions by Date}
\usage{
get_recent_late_contributions_date(cycle = 2018, year, month, day, myAPI_Key)
}
\arguments{
\item{cycle}{The election cycle}

\item{year}{The four-digit year from 2001-2018}

\item{month}{The two-digit month from 01-12}

\item{day}{The two-digit day from 01-31}
}
\description{
HTTP Request: GET https://api.propublica.org/campaign-finance/v1/{cycle}/contributions/48hour/{year}/{month}/{day}
}
\examples{
\dontrun{
get_recent_late_contributions_date(2016, year = 2015, month = '09', day = '07')
}
}
