% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recent_lobbying_representation_filings.R
\name{get_recent_lobbying_representation_filings}
\alias{get_recent_lobbying_representation_filings}
\title{Get Recent Lobbying Representation filings}
\usage{
get_recent_lobbying_representation_filings(page = 1, myAPI_Key)
}
\arguments{
\item{page}{Results come in batches of 20. The page parameter specifies which one to retrieve.}

\item{myAPI_Key}{API key from Propublica To use the congress API, you must sign up for an API key.}
}
\value{
List of returned JSON from endpoint that retrieves the 20 most recent lobbying representation filings.
}
\description{
To get the 20 most recent lobbying representation filings, use the following function. HTTP Request: GET https://api.propublica.org/congress/v1/lobbying/latest.json
}
\examples{
\donttest{
get_recent_lobbying_representation_filings()
}
}
