% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recent_official_communications_by_date.R
\name{get_recent_official_communications_by_date}
\alias{get_recent_official_communications_by_date}
\title{Get Recent Official Communications by Date}
\usage{
get_recent_official_communications_by_date(
  year,
  month,
  day,
  page = 1,
  myAPI_Key
)
}
\arguments{
\item{year}{yyyy format}

\item{month}{mm format}

\item{day}{dd format}

\item{page}{Results come in batches of 20. The page parameter specifies which one to retrieve.}

\item{myAPI_Key}{API key from Propublica To use the congress API, you must sign up for an API key.}
}
\value{
List of returned JSON from endpoint that retrieves the 20 most recent results for the specified type: ec (Executive Communication), pm (Presidential Message) or pom (Petition or Memorial).
}
\description{
o get lists of official communications to Congress on a specific date, use the following function. HTTP Request: GET https://api.propublica.org/congress/v1/{congress}/communications/category/{category}.json
}
\examples{
\donttest{
get_recent_official_communications_by_date(2018, '03', '21')
}
}
