% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recently_candidates.R
\name{get_recently_candidates}
\alias{get_recently_candidates}
\title{Get Top 20 Candidates in Specific Financial Category}
\usage{
get_recently_candidates(cycle = 2018, myAPI_Key)
}
\arguments{
\item{cycle}{The election cycle: Four-digit even-numbered year between 2010 and 2018}

\item{myAPI_Key}{use the Campaign Finance API, you must sign up for an API key. The API key must be included in all API requests to the server, set as a header.}
}
\value{
List of returned JSON from endpoint that retrieves ithe 20 most recently added FEC candidates for a given campaign cycle from the campaign-finance API from ProPublica.
}
\description{
HTTP Request
GET https://api.propublica.org/campaign-finance/v1/{cycle}/candidates/new
}
\examples{
\dontrun{
get_recently_candidates(2016)
}
}
