% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recently_committees.R
\name{get_recently_committees}
\alias{get_recently_committees}
\title{Get Recently Added Committees.}
\usage{
get_recently_committees(cycle = 2018, myAPI_Key)
}
\arguments{
\item{cycle}{The election cycle}

\item{myAPI_Key}{To use the Campaign Finance API, you must sign up for an API key. The API key must be included in all API requests to the server, set as a header.}
}
\value{
List of returned JSON from endpoint that retrieves the 20 most recently added FEC committees from the campaign-finance API from ProPublica.
}
\description{
HTTP Request: GET https://api.propublica.org/campaign-finance/v1/{cycle}/committees/new
}
\examples{
\dontrun{
get_recently_committees(2016)
}
}
