% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_senate_nomination_votes.R
\name{get_senate_nomination_votes}
\alias{get_senate_nomination_votes}
\title{Get Senate Nomination Votes}
\usage{
get_senate_nomination_votes(congress, page = 1, myAPI_Key)
}
\arguments{
\item{congress}{101-116}

\item{page}{Results come in batches of 20. The page parameter specifies which one to retrieve.}

\item{myAPI_Key}{API key from Propublica To use the congress API, you must sign up for an API key.}
}
\value{
List of returned JSON from endpoint that retrieves Senate votes on presidential nominations
}
\description{
To get Senate votes on presidential nominations. HTTP Request: GET https://api.propublica.org/congress/v1/{congress}/nominations.json
}
\examples{
\donttest{
get_senate_nomination_votes(114)
}
}
