% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_specific_lobbying_representation_filings.R
\name{get_specific_lobbying_representation_filings}
\alias{get_specific_lobbying_representation_filings}
\title{Get a Specific Lobbying Representation filing}
\usage{
get_specific_lobbying_representation_filings(id, page = 1, myAPI_Key)
}
\arguments{
\item{id}{numeric id attribute from search or latest responses}

\item{page}{Results come in batches of 20. The page parameter specifies which one to retrieve.}

\item{myAPI_Key}{API key from Propublica To use the congress API, you must sign up for an API key.}
}
\value{
List of returned JSON from endpoint that retrieves
}
\description{
To get a specific lobbying representation filing, use the following function. HTTP Request: GET https://api.propublica.org/congress/v1/lobbying/{id}.json
}
\examples{
\donttest{
get_specific_lobbying_representation_filings('300986302')
}
}
