% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_specific_subcommittee.R
\name{get_specific_subcommittee}
\alias{get_specific_subcommittee}
\title{Get a Specific Subcommittee}
\usage{
get_specific_subcommittee(
  congress,
  chamber,
  committee_id,
  subcommittee_id,
  page = 1,
  myAPI_Key
)
}
\arguments{
\item{congress}{114-116}

\item{chamber}{house or senate or joint}

\item{committee_id}{Optional committee abbreviation, for example HSAG. Use the full committees response to find abbreviations.}

\item{subcommittee_id}{Optional sub-committee abbreviation, for example HSAG. Use the full committees response to find abbreviations.}

\item{page}{Results come in batches of 20. The page parameter specifies which one to retrieve.}

\item{myAPI_Key}{API key from Propublica To use the congress API, you must sign up for an API key.}
}
\value{
List of returned JSON from endpoint that retrieves the 20 most recent hearings and supports pagination.
}
\description{
To get information about a single Senate or House subcommittee, including the members of that subcommittee, use the following function. HTTP Request: GET https://api.propublica.org/congress/v1/{congress}/{chamber}/committees/{committee-id}/subcommittees/{subcommittee-id}.json
}
\examples{
\donttest{
get_specific_subcommittee(115, 'house', 'HSAS', 'HSAS28')
get_specific_subcommittee(115, 'house', 
lists_of_committees(115, "senate")$results[[1]]$committees[[2]]$id,
lists_of_committees(115, "senate")$results[[1]]$committees[[2]]$subcommittees[[1]]$id)
}
}
