% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_state_party_counts.R
\name{get_state_party_counts}
\alias{get_state_party_counts}
\title{Get State Party Counts}
\usage{
get_state_party_counts(page = 1, myAPI_Key)
}
\arguments{
\item{page}{Results come in batches of 20. The page parameter specifies which one to retrieve.}

\item{myAPI_Key}{API key from Propublica To use the congress API, you must sign up for an API key.}
}
\value{
List of returned JSON from endpoint that retrieves
}
\description{
To get party membership counts for all states (current Congress only), use the following function. HTTP Request: GET https://api.propublica.org/congress/v1/states/members/party.json
}
\examples{
\donttest{
get_state_party_counts()
}
}
