% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary_presidential_electronic_filing.R
\name{get_summary_presidential_electronic_filing}
\alias{get_summary_presidential_electronic_filing}
\title{Get Summary for a Specific Presidential Electronic Filing}
\usage{
get_summary_presidential_electronic_filing(cycle, filing_id, myAPI_Key)
}
\arguments{
\item{cycle}{The election cycle}

\item{filing_id}{Integer representing the ID of a Form 3 electronic filing.}

\item{myAPI_Key}{use the Campaign Finance API, you must sign up for an API key. The API key must be included in all API requests to the server, set as a header.}
}
\value{
List of returned JSON from endpoint that retrieves a summary figures from a specific FEC electronic filing (Form 3 filings by presidential candidates only).
}
\description{
HTTP Request: GET https://api.propublica.org/campaign-finance/v1/{cycle}/filings/{filing_id}
}
\examples{
\dontrun{
get_summary_presidential_electronic_filing(2016, 1029629)
}
}
