% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists_of_committees.R
\name{lists_of_committees}
\alias{lists_of_committees}
\title{Lists of Committees}
\usage{
lists_of_committees(congress, chamber, page = 1, myAPI_Key)
}
\arguments{
\item{congress}{110-116}

\item{chamber}{house, senate or joint}

\item{page}{Results come in batches of 20. The page parameter specifies which one to retrieve.}

\item{myAPI_Key}{API key from Propublica To use the congress API, you must sign up for an API key.}
}
\value{
List of returned JSON from endpoint that retrieves the 20 most recent results and supports pagination using multiples of 20.
}
\description{
To get a list of Senate, House or joint committees, including their subcommittees, use the following function. HTTP Request: GET https://api.propublica.org/congress/v1/{congress}/{chamber}/committees.json
}
\examples{
\donttest{
lists_of_committees(115, "senate")
}
}
