% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/office_expenses_category.R
\name{office_expenses_category}
\alias{office_expenses_category}
\title{Get Quarterly Office Expenses for a Specified Category}
\usage{
office_expenses_category(category, year, quarter, myAPI_Key)
}
\arguments{
\item{category}{travel or personnel or rent-utilities or other-services or supplies or franked-mail or printing or equipment or total}

\item{year}{2009-2017}

\item{quarter}{1,2,3,4}

\item{myAPI_Key}{use the congress API, you must sign up for an API key. The API key must be included in all API requests to the server, set as a header.}
}
\value{
The API returns the first 20 results ordered by amount descending and supports pagination using an offset URI parameter set to multiples of 20. The response also includes the change in spending from the previous quarter.
}
\description{
The House of Representatives publishes quarterly reports detailing official office expenses by lawmakers. The Congress API has data beginning in the third quarter of 2009. To get the amount spent by individual lawmakers in a specified category during a specified year and quarter by category, use the following function. HTTP Request: GET https://api.propublica.org/congress/v1/office_expenses/category/{category}/{year}/{quarter}.json
}
\examples{
\dontrun{
office_expenses_category('travel', 2017, 4)
}
}
