% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_key.R
\name{retrieve_key}
\alias{retrieve_key}
\title{Retrieve the relevant ProPublica API key for congress API or campaign finance API}
\usage{
retrieve_key(API = c("congress", "campaign-finance"))
}
\arguments{
\item{API}{Which API are you accessing}
}
\value{
the key from config.yml or global environment variables key_congress/key_campaign_finance
}
\description{
This function is primarily meant to be used by pp_query to determine the API keys if not passed explicitly.
}
\examples{
\donttest{
retrieve_key('congress')
retrieve_key('campaign-finance')
}
}
