% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{us_congressional_districts}
\alias{us_congressional_districts}
\title{US Congressional Districts}
\format{
A data frame with 435 rows and 3 variables:
\describe{
\item{district}{The district name}
\item{state_name}{The state the district is in}
\item{state_abb}{Two letter state abbreviation}
\item{district_number}{District number. Except for at-large districts.}
...
}
}
\usage{
us_congressional_districts
}
\description{
A dataset containing some information about US
Congressional Districts
}
\keyword{datasets}
