% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_cycle.R
\name{validate_cycle}
\alias{validate_cycle}
\title{Validate that year is correct}
\usage{
validate_cycle(cycle)
}
\arguments{
\item{cycle}{}
}
\value{
TRUE if correct FALSE otherwise
}
\description{
Validate that year is correct
}
\examples{
validate_cycle(1979)#TRUE
validate_cycle(1979)#FALSE
validate_cycle(16) #FALSE
}
