% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_district.R
\name{validate_district}
\alias{validate_district}
\title{Validate that a district abbreviation is correct}
\usage{
validate_district(us_district)
}
\arguments{
\item{us_district}{An integer corresponding to the district.}
}
\value{
TRUE if correct FALSE otherwise
}
\description{
Validate that a district abbreviation is correct
}
\examples{
\donttest{
validate_district(1)#TRUE
validate_district(111) #FALSE
validate_district(us_congressional_districts$district[1]) #FALSE
validate_district(as.numeric(us_congressional_districts$district_number[1])) #TRUE
}
}
