% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_candidate.R
\name{get_candidate}
\alias{get_candidate}
\title{Get Candidate Info}
\usage{
get_candidate(FEC_ID, cycle = 2018)
}
\arguments{
\item{FEC_ID}{The FEC-assigned 9-character ID of a candidate.}

\item{cycle}{The election cycle}
}
\description{
Get Candidate Info
}
\examples{
\dontrun{
get_candidate('P60005915', 2016)
}
}
