% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_candidate.R
\name{get_candidate}
\alias{get_candidate}
\title{Get Candidate Info
https://www.propublica.org/datastore/apis
GET https://api.propublica.org/campaign-finance/v1/{cycle}/candidates/{fec-id}}
\usage{
get_candidate(FEC_ID, cycle = 2018, page = 1, myAPI_Key)
}
\arguments{
\item{FEC_ID}{The FEC-assigned 9-character ID of a candidate.}

\item{cycle}{The election cycle must be even-numbered year between 1996 and 2018}

\item{page}{Results come in batches of 20. The page parameter specifies which one to retrieve.}

\item{myAPI_Key}{API key from Propublica To use the congress API, you must sign up for an API key.}
}
\value{
a list object with the return values about candidate basic information
}
\description{
Get Candidate Info
https://www.propublica.org/datastore/apis
GET https://api.propublica.org/campaign-finance/v1/{cycle}/candidates/{fec-id}
}
\examples{
\donttest{
get_candidate('P60005915', 2016)}

}
