% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_candidate_by_name.R
\name{get_candidate_by_name}
\alias{get_candidate_by_name}
\title{Get Candidate Info}
\usage{
get_candidate_by_name(NAME, cycle = 2018, myAPI_Key)
}
\arguments{
\item{NAME}{The first or last name of the candidate.}

\item{cycle}{The election cycle}
}
\description{
HTTP Request: GET https://api.propublica.org/campaign-finance/v1/{cycle}/candidates/search
}
\examples{
get_candidate_by_name('Wilson', 2016)

}
