% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_get_new_members.R
\name{get_new_members}
\alias{get_new_members}
\title{Get New Members}
\usage{
get_new_members(myAPI_Key)
}
\arguments{
\item{myAPI_Key}{API key from Propublica To use the congress API, you must sign up for an API key.}
}
\value{
List of returned JSON from endpoint that retrieves a list of the most recent new members of the current Congress
}
\description{
To get a list of the most recent new members of the current Congress. 
HTTP Request: GET https://api.propublica.org/congress/v1/members/new.json
}
\examples{
\donttest{
get_new_members()
}
}
