% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recent_amendments.R
\name{get_recent_amendments}
\alias{get_recent_amendments}
\title{Get Recent Amendments}
\usage{
get_recent_amendments(cycle = 2018, myAPI_Key)
}
\arguments{
\item{cycle}{The election cycle}

\item{myAPI_Key}{API key from Propublica To use the congress API, you must sign up for an API key.}
}
\value{
List of returned JSON from endpoint that retrieves the most recent filings that are amendments of earlier filings
}
\description{
HTTP Request: GET https://api.propublica.org/campaign-finance/v1/{cycle}/filings/amendments
}
\examples{
\donttest{
get_recent_amendments(2016)
}
}
