% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recent_late_contributions.R
\name{get_recent_late_contributions}
\alias{get_recent_late_contributions}
\title{Get Recent Late Contributions
https://www.propublica.org/datastore/apis
HTTP Request: GET https://api.propublica.org/campaign-finance/v1/{cycle}/contributions/48hour}
\usage{
get_recent_late_contributions(cycle = 2018, page = 1, myAPI_Key)
}
\arguments{
\item{cycle}{The election cycle. Cycle must be even-numbered year between 2010 and 2018}

\item{page}{Results come in batches of 20. The page parameter specifies which one to retrieve.}

\item{myAPI_Key}{API key from Propublica To use the congress API, you must sign up for an API key.}
}
\value{
List of returned JSON from endpoint that retrieves the most recent late contributions to candidates from the campaign-finance API from ProPublica.
}
\description{
Get Recent Late Contributions
https://www.propublica.org/datastore/apis
HTTP Request: GET https://api.propublica.org/campaign-finance/v1/{cycle}/contributions/48hour
}
\examples{
\donttest{
get_recent_late_contributions(2016)
}
}
