% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recent_late_contributions_committee.R
\name{get_recent_late_contributions_committee}
\alias{get_recent_late_contributions_committee}
\title{Get Recent Late Contributions to a Specific Committee
https://www.propublica.org/datastore/apis
HTTP Request: GET https://api.propublica.org/campaign-finance/v1/{cycle}/committees/{fec-id}/48hour}
\usage{
get_recent_late_contributions_committee(
  cycle = 2018,
  FEC_ID,
  page = 1,
  myAPI_Key
)
}
\arguments{
\item{cycle}{The election cycle}

\item{FEC_ID}{The FEC-assigned 9-character ID of a committee.}

\item{page}{Results come in batches of 20. The page parameter specifies which one to retrieve.}

\item{myAPI_Key}{API key from Propublica To use the congress API, you must sign up for an API key.}
}
\value{
List of returned JSON from endpoint that retrieves the most recent late contributions to a specific committee.
}
\description{
Get Recent Late Contributions to a Specific Committee
https://www.propublica.org/datastore/apis
HTTP Request: GET https://api.propublica.org/campaign-finance/v1/{cycle}/committees/{fec-id}/48hour
}
\examples{
\donttest{
get_recent_late_contributions_committee(2016, 'C00575050')
}
}
