% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_specific_nominations.R
\name{get_specific_nominations}
\alias{get_specific_nominations}
\title{Get a Specific Nomination}
\usage{
get_specific_nominations(congress, nominee_id, page = 1, myAPI_Key)
}
\arguments{
\item{congress}{107-116}

\item{nominee_id}{alphanumeric ID beginning with PN - for example, PN675}

\item{page}{Results come in batches of 20. The page parameter specifies which one to retrieve.}

\item{myAPI_Key}{API key from Propublica To use the congress API, you must sign up for an API key.}
}
\value{
List of returned JSON from endpoint that retrieves
}
\description{
To get details about a particular presidential civilian nomination,, use the following function. HTTP Request: GET https://api.propublica.org/congress/v1/{congress}/nominees/{nominee-id}.json
}
\examples{
\donttest{
get_specific_nominations(115, 'PN40')
}
}
