% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_cycle.R
\name{validate_cycle}
\alias{validate_cycle}
\title{Validate that year cycle is correct}
\usage{
validate_cycle(cycle)
}
\arguments{
\item{cycle}{An election cycle (e.g. 2018)}
}
\value{
TRUE if correct FALSE otherwise
}
\description{
Validate that year cycle is correct
}
\examples{
validate_cycle(1996)#TRUE
validate_cycle(1979)#FALSE
validate_cycle(16) #FALSE
}
